Refactor camera processing scenario. Move indexed paths to main script, and refactor export function.
