+++ auvtools/camera/__init__.py
@@ -1,2 +1,2 @@
 from .camera import Geolocation, Position3D, Orientation3D, ImageFile, Camera
-from .file_index import FileIndex
+from .writer import write_cameras_to_file
